Refactor NPC shop creation and management logic
Fixed the issue where the prompt for The Wandering Shop didn’t appear at startup

+++ rs_stores/client.lua
@@ -1,30 +1,125 @@
-local rutaActual = 1
-local npc = nil
 local openmenu = nil
-local npcBlip = nil
 local tiendaPrompts = {}
 local tiendaBlips = {}
-local NpcShop = {}
+local NpcTienda = {}
+local npcshop = nil
+local npcBlip = nil
+local rutaActual = 1
+local showingPrompt = false
+local npcshop_creating = false
 
 Citizen.CreateThread(function()
-    local showingPrompt = false
+    while not Config or not Config.rutas or #Config.rutas == 0 do
+        Wait(1000)
+    end
 
+    Wait(1000)
+
+    if Config.TiendaNPCActiva then
+        crearNPCShop()
+    end
+end)
+
+function crearNPCShop()
+    if npcshop_creating then
+        return
+    end
+    npcshop_creating = true
+
+    if npcshop and DoesEntityExist(npcshop) then
+        DeleteEntity(npcshop)
+        npcshop = nil
+    end
+
+    local model = GetHashKey(Config.npcModel)
+    if not IsModelValid(model) then
+        npcshop_creating = false
+        return
+    end
+
+    RequestModel(model)
+    while not HasModelLoaded(model) do Wait(10) end
+
+    local datos = Config.rutas[rutaActual]
+    if not datos or not datos.coords then
+        npcshop_creating = false
+        return
+    end
+
+    local x, y, z = datos.coords.x, datos.coords.y, datos.coords.z - 1.0
+    local heading = datos.heading or 0.0
+
+    npcshop = Citizen.InvokeNative(0xD49F9B0955C367DE, model, x, y, z, heading, true, true, true, true)
+
+    if npcshop and DoesEntityExist(npcshop) then
+        Citizen.InvokeNative(0x283978A15512B2FE, npcshop, true)
+        SetEntityNoCollisionEntity(PlayerPedId(), npcshop, true)
+        SetEntityCanBeDamaged(npcshop, false)
+        SetEntityInvincible(npcshop, true)
+        FreezeEntityPosition(npcshop, true)
+        SetBlockingOfNonTemporaryEvents(npcshop, true)
+        SetModelAsNoLongerNeeded(model)
+
+        if Config.TiendaBlipActivo then
+            if npcBlip then RemoveBlip(npcBlip) end
+            local coords = GetEntityCoords(npcshop)
+            npcBlip = Citizen.InvokeNative(0x554D9D53F696D002, 1664425300, coords.x, coords.y, coords.z)
+            SetBlipSprite(npcBlip, Config.spriteBlip, true)
+            SetBlipScale(npcBlip, 0.8)
+            Citizen.InvokeNative(0x9CB1A1623062F402, npcBlip, Config.nameBlip)
+        end
+
+        npcshop_creating = false
+        moverNPCShop()
+    else
+        npcshop_creating = false
+    end
+end
+
+Citizen.CreateThread(function()
+    while true do
+        Wait(5000)
+        if Config.TiendaNPCActiva then
+            if not npcshop_creating and (not npcshop or not DoesEntityExist(npcshop)) then
+                crearNPCShop()
+            end
+        end
+    end
+end)
+
+function moverNPCShop()
+    Citizen.CreateThread(function()
+        while Config.TiendaNPCActiva and npcshop and DoesEntityExist(npcshop) do
+            local destino = Config.rutas[rutaActual]
+            if destino then
+                local x, y, z = destino.coords.x, destino.coords.y, destino.coords.z - 1.0
+                local heading = destino.heading or 0.0
+                SetEntityCoords(npcshop, x, y, z, false, false, false, true)
+                SetEntityHeading(npcshop, heading)
+                FreezeEntityPosition(npcshop, true)
+
+                if npcBlip then SetBlipCoords(npcBlip, x, y, z) end
+            end
+
+            Citizen.Wait(Config.merchattimelocation * 60 * 1000)
+            rutaActual = rutaActual % #Config.rutas + 1
+            FreezeEntityPosition(npcshop, false)
+        end
+    end)
+end
+
+Citizen.CreateThread(function()
     while true do
         local playerPed = PlayerPedId()
         local playerCoords = GetEntityCoords(playerPed)
         local nearNPC = false
 
-        if Config.TiendaNPCActiva and npc and DoesEntityExist(npc) then
-            local npcCoords = GetEntityCoords(npc)
-
+        if Config.TiendaNPCActiva and npcshop and DoesEntityExist(npcshop) then
+            local npcCoords = GetEntityCoords(npcshop)
             if #(playerCoords - npcCoords) < 3.0 then
                 nearNPC = true
-
                 if not showingPrompt then
-                    SendNUIMessage({
-                        type = "showmerchant",
-                        text = Config.promptmerchant
-                    })
+                    SendNUIMessage({ type = "showmerchant", text = Config.promptmerchant })
                     showingPrompt = true
                 end
 
@@ -42,63 +137,6 @@ Citizen.CreateThread(function()
         Citizen.Wait(nearNPC and 0 or 500)
     end
 end)
-
-if Config.TiendaNPCActiva and (not npc or not DoesEntityExist(npc)) then
-    Citizen.CreateThread(function()
-        local model = GetHashKey(Config.npcModel)
-        RequestModel(model)
-        while not HasModelLoaded(model) do Wait(1) end
-
-        local datos = Config.rutas[rutaActual]
-        local x, y, z = datos.coords.x, datos.coords.y, datos.coords.z - 1.0
-        local heading = datos.heading or 0.0
-
-        npc = Citizen.InvokeNative(0xD49F9B0955C367DE, model, x, y, z, heading, true, true, true, true)
-
-        if npc and DoesEntityExist(npc) then
-            Citizen.InvokeNative(0x283978A15512B2FE, npc, true)
-            SetEntityNoCollisionEntity(PlayerPedId(), npc, true)
-            SetEntityCanBeDamaged(npc, false)
-            SetEntityInvincible(npc, true)
-            FreezeEntityPosition(npc, true)
-            SetBlockingOfNonTemporaryEvents(npc, true)
-            SetModelAsNoLongerNeeded(model)
-
-            if Config.TiendaBlipActivo then
-                local coords = GetEntityCoords(npc)
-                npcBlip = Citizen.InvokeNative(0x554D9D53F696D002, 1664425300, coords.x, coords.y, coords.z)
-                SetBlipSprite(npcBlip, Config.spriteBlip, true)
-                SetBlipScale(npcBlip, 0.8)
-                Citizen.InvokeNative(0x9CB1A1623062F402, npcBlip, Config.nameBlip)
-            end
-
-            moverNPC()
-        end
-    end)
-end
-
-function moverNPC()
-    Citizen.CreateThread(function()
-        while true do
-            local destino = Config.rutas[rutaActual]
-            local x, y, z = destino.coords.x, destino.coords.y, destino.coords.z - 1.0
-            local heading = destino.heading or 0.0
-
-            SetEntityCoords(npc, x, y, z, false, false, false, true)
-            SetEntityHeading(npc, heading)
-            FreezeEntityPosition(npc, true)
-
-            if npcBlip then
-                SetBlipCoords(npcBlip, x, y, z)
-            end
-
-            Citizen.Wait(Config.merchattimelocation * 60 * 1000)
-
-            rutaActual = rutaActual % #Config.rutas + 1
-            FreezeEntityPosition(npc, false)
-        end
-    end)
-end
 
 function CrearPromptTienda(id, label, coords)
     local groupId = GetRandomIntInRange(0, 0xffffff)
@@ -143,7 +181,6 @@ Citizen.CreateThread(function()
     while true do
         Citizen.Wait(0)
         local playerCoords = GetEntityCoords(PlayerPedId())
-
         for id, datos in pairs(tiendaPrompts) do
             if #(playerCoords - datos.coords) < 2.5 then
                 UiPromptSetActiveGroupThisFrame(datos.group, datos.label)
@@ -233,34 +270,22 @@ AddEventHandler("rs_stores:CreateNPC", function(coords, modelName)
     FreezeEntityPosition(npc, true)
     SetBlockingOfNonTemporaryEvents(npc, true)
     SetModelAsNoLongerNeeded(model)
-
-    table.insert(NpcShop, npc)
+    table.insert(NpcTienda, npc)
 end)
 
 AddEventHandler("onResourceStop", function(resourceName)
     if resourceName == GetCurrentResourceName() then
-        if npc and DoesEntityExist(npc) then
-            DeleteEntity(npc)
-            npc = nil
+        if npcshop and DoesEntityExist(npcshop) then DeleteEntity(npcshop) end
+        if npcBlip then RemoveBlip(npcBlip) end
+
+        for _, blip in pairs(tiendaBlips) do
+            if blip then RemoveBlip(blip) end
         end
 
-        if npcBlip then
-            RemoveBlip(npcBlip)
-            npcBlip = nil
+        for _, entidad in pairs(NpcTienda) do
+            if entidad and DoesEntityExist(entidad) then DeleteEntity(entidad) end
         end
 
-        for nombre, blip in pairs(tiendaBlips) do
-            if blip then
-                RemoveBlip(blip)
-            end
-        end
-        tiendaBlips = {}
-
-        for _, entidad in pairs(NpcShop) do
-            if entidad and DoesEntityExist(entidad) then
-                DeleteEntity(entidad)
-            end
-        end
-        NpcShop = {}
+        npcshop, npcBlip, tiendaBlips, NpcTienda = nil, nil, {}, {}
     end
 end)
